# packages/feedpon-state/openapi/feedly.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Feedly API",
    "version": "3"
  },
  "servers": [
    {
      "url": "https://cloud.feedly.com/v3"
    },
    {
      "url": "https://sandbox7.feedly.com"
    }
  ],
  "security": [
    {
      "OAuth2": ["https://cloud.feedly.com/subscriptions"]
    }
  ],
  "paths": {
    "/auth/{exchangeToken}": {
      "post": {
        "summary": "Exchanging an auth code for a refresh token and an access token",
        "operationId": "exchangeToken",
        "security": [],
        "parameters": [
          {
            "name": "exchangeToken",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["token"]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code",
                  "client_id",
                  "client_secret",
                  "redirect_uri",
                  "grant_type"
                ],
                "properties": {
                  "code": {
                    "type": "string",
                    "description": "the code returned from the previous call."
                  },
                  "client_id": {
                    "type": "string",
                    "description": "the clientId obtained during application registration."
                  },
                  "client_secret": {
                    "type": "string",
                    "description": "the client secret obtained during application registration."
                  },
                  "redirect_uri": {
                    "type": "string",
                    "description": "the URI registered with the application (if you pass it as a URL parameter, be sure to URL-encode it!)."
                  },
                  "state": {
                    "type": "string",
                    "description": "indicates any state which may be useful to your application upon receipt of the response. The feedly Authorization Server roundtrips this parameter, so your application receives the same value it sent. Possible uses include redirecting the user to the correct resource in your site, nonces, and cross-site-request-forgery mitigations."
                  },
                  "grant_type": {
                    "type": "string",
                    "enum": ["authorization_code"],
                    "description": "as defined in the OAuth2 specification, this field must be set to `authorization_code`."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "access_token",
                    "refresh_token",
                    "expires_in",
                    "token_type",
                    "plan"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "the feedly user id"
                    },
                    "access_token": {
                      "type": "string",
                      "description": "a token that may be used to access APIs. Access tokens are have an expiration"
                    },
                    "refresh_token": {
                      "type": "string",
                      "description": "a token that may be used to obtain a new access token. Refresh tokens are valid until the user revokes access."
                    },
                    "expires_in": {
                      "type": "integer",
                      "description": "the remaining lifetime on the access token"
                    },
                    "token_type": {
                      "type": "string",
                      "description": "indicates the type of token returned. At this time, this field will always have the value of Bearer"
                    },
                    "plan": {
                      "type": "string",
                      "enum": ["standard", "pro", "business"],
                      "description": "indicated the user plan"
                    },
                    "state": {
                      "type": "string",
                      "description": "the state that was passed in"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        }
      }
    },
    "/auth/{refreshToken}": {
      "post": {
        "summary": "Refreshing an access token",
        "operationId": "refreshToken",
        "security": [],
        "parameters": [
          {
            "name": "refreshToken",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["token"]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "refresh_token",
                  "client_id",
                  "client_secret",
                  "grant_type"
                ],
                "properties": {
                  "refresh_token": {
                    "type": "string",
                    "description": "the refresh token returned in the previous code."
                  },
                  "client_id": {
                    "type": "string",
                    "description": "the clientId obtained during application registration."
                  },
                  "client_secret": {
                    "type": "string",
                    "description": "the client secret obtained during application registration."
                  },
                  "grant_type": {
                    "type": "string",
                    "enum": ["refresh_token"],
                    "description": "as defined in the OAuth2 specification, this field must be set to `refresh_token`."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "access_token",
                    "expires_in",
                    "token_type",
                    "plan"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "the feedly user id"
                    },
                    "access_token": {
                      "type": "string",
                      "description": "a token that may be used to access APIs. Access tokens are have an expiration"
                    },
                    "expires_in": {
                      "type": "integer",
                      "description": "the remaining lifetime on the access token"
                    },
                    "token_type": {
                      "type": "string",
                      "description": "indicates the type of token returned. At this time, this field will always have the value of Bearer"
                    },
                    "plan": {
                      "type": "string",
                      "enum": ["standard", "pro", "business"],
                      "description": "indicated the user plan"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "UnexpectedError": {
        "description": "Unexpected error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "text/plain": {
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "OAuth2": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "/auth/auth",
            "scopes": {
              "https://cloud.feedly.com/subscriptions": "grant all access"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}